refactor(observatility): move tracing and logs in observability
This commit moves the tracing and logging core code in the observability
folder.
