+++ kong-3.8.0-0.rockspec
@@ -624,29 +624,29 @@ build = {
     ["kong.vaults.env"] = "kong/vaults/env/init.lua",
     ["kong.vaults.env.schema"] = "kong/vaults/env/schema.lua",
 
-    ["kong.tracing.instrumentation"] = "kong/tracing/instrumentation.lua",
-    ["kong.tracing.propagation"] = "kong/tracing/propagation/init.lua",
-    ["kong.tracing.propagation.schema"] = "kong/tracing/propagation/schema.lua",
-    ["kong.tracing.propagation.utils"] = "kong/tracing/propagation/utils.lua",
-    ["kong.tracing.propagation.extractors._base"] = "kong/tracing/propagation/extractors/_base.lua",
-    ["kong.tracing.propagation.extractors.w3c"] = "kong/tracing/propagation/extractors/w3c.lua",
-    ["kong.tracing.propagation.extractors.b3"] = "kong/tracing/propagation/extractors/b3.lua",
-    ["kong.tracing.propagation.extractors.jaeger"] = "kong/tracing/propagation/extractors/jaeger.lua",
-    ["kong.tracing.propagation.extractors.ot"] = "kong/tracing/propagation/extractors/ot.lua",
-    ["kong.tracing.propagation.extractors.gcp"] = "kong/tracing/propagation/extractors/gcp.lua",
-    ["kong.tracing.propagation.extractors.aws"] = "kong/tracing/propagation/extractors/aws.lua",
-    ["kong.tracing.propagation.extractors.datadog"] = "kong/tracing/propagation/extractors/datadog.lua",
-    ["kong.tracing.propagation.injectors._base"] = "kong/tracing/propagation/injectors/_base.lua",
-    ["kong.tracing.propagation.injectors.w3c"] = "kong/tracing/propagation/injectors/w3c.lua",
-    ["kong.tracing.propagation.injectors.b3"] = "kong/tracing/propagation/injectors/b3.lua",
-    ["kong.tracing.propagation.injectors.b3-single"] = "kong/tracing/propagation/injectors/b3-single.lua",
-    ["kong.tracing.propagation.injectors.jaeger"] = "kong/tracing/propagation/injectors/jaeger.lua",
-    ["kong.tracing.propagation.injectors.ot"] = "kong/tracing/propagation/injectors/ot.lua",
-    ["kong.tracing.propagation.injectors.gcp"] = "kong/tracing/propagation/injectors/gcp.lua",
-    ["kong.tracing.propagation.injectors.aws"] = "kong/tracing/propagation/injectors/aws.lua",
-    ["kong.tracing.propagation.injectors.datadog"] = "kong/tracing/propagation/injectors/datadog.lua",
-    ["kong.tracing.request_id"] = "kong/tracing/request_id.lua",
-    ["kong.tracing.tracing_context"] = "kong/tracing/tracing_context.lua",
+    ["kong.observability.tracing.instrumentation"] = "kong/observability/tracing/instrumentation.lua",
+    ["kong.observability.tracing.propagation"] = "kong/observability/tracing/propagation/init.lua",
+    ["kong.observability.tracing.propagation.schema"] = "kong/observability/tracing/propagation/schema.lua",
+    ["kong.observability.tracing.propagation.utils"] = "kong/observability/tracing/propagation/utils.lua",
+    ["kong.observability.tracing.propagation.extractors._base"] = "kong/observability/tracing/propagation/extractors/_base.lua",
+    ["kong.observability.tracing.propagation.extractors.w3c"] = "kong/observability/tracing/propagation/extractors/w3c.lua",
+    ["kong.observability.tracing.propagation.extractors.b3"] = "kong/observability/tracing/propagation/extractors/b3.lua",
+    ["kong.observability.tracing.propagation.extractors.jaeger"] = "kong/observability/tracing/propagation/extractors/jaeger.lua",
+    ["kong.observability.tracing.propagation.extractors.ot"] = "kong/observability/tracing/propagation/extractors/ot.lua",
+    ["kong.observability.tracing.propagation.extractors.gcp"] = "kong/observability/tracing/propagation/extractors/gcp.lua",
+    ["kong.observability.tracing.propagation.extractors.aws"] = "kong/observability/tracing/propagation/extractors/aws.lua",
+    ["kong.observability.tracing.propagation.extractors.datadog"] = "kong/observability/tracing/propagation/extractors/datadog.lua",
+    ["kong.observability.tracing.propagation.injectors._base"] = "kong/observability/tracing/propagation/injectors/_base.lua",
+    ["kong.observability.tracing.propagation.injectors.w3c"] = "kong/observability/tracing/propagation/injectors/w3c.lua",
+    ["kong.observability.tracing.propagation.injectors.b3"] = "kong/observability/tracing/propagation/injectors/b3.lua",
+    ["kong.observability.tracing.propagation.injectors.b3-single"] = "kong/observability/tracing/propagation/injectors/b3-single.lua",
+    ["kong.observability.tracing.propagation.injectors.jaeger"] = "kong/observability/tracing/propagation/injectors/jaeger.lua",
+    ["kong.observability.tracing.propagation.injectors.ot"] = "kong/observability/tracing/propagation/injectors/ot.lua",
+    ["kong.observability.tracing.propagation.injectors.gcp"] = "kong/observability/tracing/propagation/injectors/gcp.lua",
+    ["kong.observability.tracing.propagation.injectors.aws"] = "kong/observability/tracing/propagation/injectors/aws.lua",
+    ["kong.observability.tracing.propagation.injectors.datadog"] = "kong/observability/tracing/propagation/injectors/datadog.lua",
+    ["kong.observability.tracing.request_id"] = "kong/observability/tracing/request_id.lua",
+    ["kong.observability.tracing.tracing_context"] = "kong/observability/tracing/tracing_context.lua",
 
     ["kong.observability.logs"] = "kong/observability/logs.lua",
