+++ kong/error_handlers.lua
@@ -1,7 +1,7 @@
 local kong = kong
 local find = string.find
 local fmt  = string.format
-local request_id = require "kong.tracing.request_id"
+local request_id = require "kong.observability.tracing.request_id"
 local tools_http = require "kong.tools.http"
