+++ kong/observability/tracing/propagation/extractors/ot.lua
@@ -1,5 +1,5 @@
-local _EXTRACTOR               = require "kong.tracing.propagation.extractors._base"
-local propagation_utils        = require "kong.tracing.propagation.utils"
+local _EXTRACTOR               = require "kong.observability.tracing.propagation.extractors._base"
+local propagation_utils        = require "kong.observability.tracing.propagation.utils"
 
 local from_hex                 = propagation_utils.from_hex
 local parse_baggage_headers    = propagation_utils.parse_baggage_headers
