+++ kong/observability/tracing/propagation/injectors/b3-single.lua
@@ -1,4 +1,4 @@
-local _INJECTOR = require "kong.tracing.propagation.injectors._base"
+local _INJECTOR = require "kong.observability.tracing.propagation.injectors._base"
 local to_hex    = require "resty.string".to_hex
 
 local B3_SINGLE_INJECTOR = _INJECTOR:new({
