+++ kong/pdk/log.lua
@@ -19,7 +19,7 @@ local constants = require("kong.constants")
 local clear_tab = require("table.clear")
 
 
-local request_id_get = require("kong.tracing.request_id").get
+local request_id_get = require("kong.observability.tracing.request_id").get
 local cycle_aware_deep_copy = require("kong.tools.table").cycle_aware_deep_copy
 local get_tls1_version_str = require("ngx.ssl").get_tls1_version_str
 local get_workspace_name = require("kong.workspaces").get_workspace_name
