+++ kong/plugins/opentelemetry/traces.lua
@@ -1,6 +1,6 @@
 local Queue = require "kong.tools.queue"
-local propagation = require "kong.tracing.propagation"
-local tracing_context = require "kong.tracing.tracing_context"
+local propagation = require "kong.observability.tracing.propagation"
+local tracing_context = require "kong.observability.tracing.tracing_context"
 local otlp = require "kong.plugins.opentelemetry.otlp"
 local otel_utils = require "kong.plugins.opentelemetry.utils"
 local clone = require "table.clone"
