+++ spec/01-unit/10-log_serializer_spec.lua
@@ -54,7 +54,7 @@ describe("kong.log.serialize", function()
         get_phase = function() return "access" end,
       }
 
-      package.loaded["kong.tracing.request_id"] = nil
+      package.loaded["kong.observability.tracing.request_id"] = nil
       package.loaded["kong.pdk.log"] = nil
       kong.log = require "kong.pdk.log".new(kong)
