+++ spec/01-unit/26-observability/03-propagation_module_spec.lua
@@ -1,4 +1,4 @@
-local propagation_utils = require "kong.tracing.propagation.utils"
+local propagation_utils = require "kong.observability.tracing.propagation.utils"
 local tablex = require "pl.tablex"
 local shallow_copy = require "kong.tools.table".shallow_copy
 local to_hex = require "resty.string".to_hex
@@ -423,7 +423,7 @@ describe("Tracing Headers Propagation Module", function()
       }
     }
   }
-  local propagation      = require "kong.tracing.propagation"
+  local propagation      = require "kong.observability.tracing.propagation"
 
   lazy_setup(function()
     err                 = spy.on(kong.log, "err")
