+++ spec/03-plugins/27-aws-lambda/05-aws-serializer_spec.lua
@@ -10,7 +10,7 @@ describe("[AWS Lambda] aws-gateway input", function()
 
   local function reload_module()
     -- make sure to reload the module
-    package.loaded["kong.tracing.request_id"] = nil
+    package.loaded["kong.observability.tracing.request_id"] = nil
     package.loaded["kong.plugins.aws-lambda.request-util"] = nil
     aws_serialize = require "kong.plugins.aws-lambda.request-util".aws_serializer
   end
